+++ CameraPanel/FuseCameraImpl.h
@@ -1,12 +1,3 @@
-/*
-     File: FuseCameraImpl.h
- Abstract: Fuse Glue.
-  Version: 1.0
- 
- Copyright (C) 2015 Bjørn-Olav Strand All Rights Reserved.
- 
- */
-
 #import <UIKit/UIKit.h>
 #import <GLKit/GLKit.h>
 #import <AVFoundation/AVFoundation.h>
@@ -15,6 +6,7 @@
  	- (int) textureHeight;
  	- (int) textureWidth;
  	- (int) getRotation;
+ 	- (void*)getStillImageOutputHandle;
  	- (void)startCam:(int)device;
  	- (void)stopCam;
  	- (void)addUpdateListener:(uDelegate *)callback;
