--- CameraStream_example.unoproj
@@ -1,13 +1,0 @@
-{
-  "Name": "CameraStreamExample",
-  "RootNamespace":"",
-  "Packages": [
-    "Fuse",
-    "FuseJS",
-    "Android"
-  ],
-  "Includes": [
-    "*",
-    "CameraAndroid.java:Java:Android"
-  ]
-}
