--- MainView.ux
@@ -1,11 +1,0 @@
-<App>
-	<ScrollView>
-	<StackPanel>
-		<Text Value="Teksten!!!" />
-		<CameraStream>
-			<CameraVisual Facing="Back" />
-		</CameraStream>
-		<Text Value="Underteksten!!!" />
-	</StackPanel>
-</ScrollView>
-</App>
